requirements.txt
Django~=5.1.3
djangorestframework==3.15.2
Pillow==11.0.0
